Revert "Inline UOM classes (#47)"
This reverts commit 7bdb9e9b7061805c7eb2a0121c49519ce3ea8b80.

+++ architecture/jvmfrc/src/main/kotlin/com/lynbrookrobotics/kapuchin/preferences/actual.kt
@@ -14,7 +14,7 @@ actual fun Named.pref(fallback: Int) = Preference(this, fallback, impl::putInt, 
 actual fun Named.pref(fallback: Long) = Preference(this, fallback, impl::putLong, impl::getLong)
 actual fun Named.pref(fallback: String) = Preference(this, fallback, impl::putString, impl::getString)
 actual fun <Q : Quan<Q>> Named.pref(fallback: Number, withUnits: UomConverter<Q>) = Preference(
-        this, withUnits(fallback.toDouble()),
+        this, withUnits(fallback),
         { name, value -> impl.putDouble(name, withUnits(value)) },
         { name, value -> withUnits(impl.getDouble(name, withUnits(value))) },
         " (${withUnits.unitName})"
